requirements.txt
fastapi
uvicorn[standard]
yfinance
pandas
python-dotenv
matplotlib
